source/requirements.txt
asgiref==3.3.1
Django==3.0
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
Pillow==7.2.0
psycopg2-binary==2.8.6
